+++ starkinfra/issuingcard/__issuingcard.py
@@ -5,7 +5,6 @@ from ..issuingrule import parse_rules
 
 
 class IssuingCard(Resource):
-
     """# IssuingCard object
     The IssuingCard object displays the information of the cards created in your Workspace.
     Sensitive information will only be returned when the "expand" parameter is used, to avoid security concerns.
@@ -121,7 +120,7 @@ def page(cursor=None, limit=None, ids=None, after=None, before=None, status=None
     Receive a list of IssuingCard objects previously created in the Stark Infra API and the cursor to the next page.
     ## Parameters (optional):
     - cursor [string, default None]: cursor returned on the previous page function call
-    - limit [integer, default 100]: maximum number of objects to be retrieved. Unlimited if None. ex: 35
+    - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
     - ids [list of strings, default None]: list of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
     - after [datetime.date or string, default None] date filter for objects created only after specified date. ex: datetime.date(2020, 3, 10)
     - before [datetime.date or string, default None] date filter for objects created only before specified date. ex: datetime.date(2020, 3, 10)